rcommit

+++ src/components/Admin/Products.jsx
@@ -5,6 +5,7 @@ import AdminHeader from "./AdminHeader";
 import Button from "@mui/material/Button";
 import Edit from "./Edit";
 import axios from "axios";
+import { Link } from "react-router-dom"
 
 const Products = () => {
   const [data, setData] = useState([]);
@@ -40,7 +41,10 @@ const Products = () => {
       return;
     }
 
-    try {
+    const shouldRemove = confirm("are you sure you want to delete?")
+
+    if (shouldRemove) {
+      console.log("test:",id)
       const res = await axios.delete(
         `http://localhost:8000/api/form/delete-product/${id}`,
         {
@@ -48,21 +52,22 @@ const Products = () => {
             "Content-Type": "application/json",
           },
         }
-      );
+      ).then((res)=>{
+        getPostData();
 
-      if (res.status === 200) {
+      }).catch((err)=>{
+
+      })
+
+     
         
-        const updatedData = data.filter(
-          (currProduct) => currProduct.id !== id
-        );
-        setData(updatedData);
-        console.log("Product deleted successfully!");
-      } else {
-        console.error(`Unexpected response: ${res.status}`);
-      }
-    } catch (error) {
-      console.error("Error deleting product:", error);
     }
+     
+       
+      
+    // } catch (error) {
+    //   console.error("Error deleting product:", error);
+    // }
   };
 
   useEffect(() => {
@@ -88,11 +93,13 @@ const Products = () => {
                       key={item.id}
                       className="bg-white shadow-lg rounded-lg p-4 flex flex-col items-center hover:shadow-2xl transition-shadow"
                     >
+                      <Link to={`/products/${item._id}`}>
                       <img
                         src={item.image}
                         alt={item.name}
                         className="h-32 w-32 object-contain mb-4"
                       />
+                      </Link>
                       <h3 className="text-lg font-semibold text-gray-800 mb-2">
                         {item.name}
                       </h3>
@@ -115,8 +122,8 @@ const Products = () => {
                         <button
                           className="bg-red-500 text-white px-4 py-2 rounded hover:bg-red-600 transition"
                           onClick={() => {
-                            console.log("Deleting item with ID:", item.id); // Debugging
-                            handleDeleteProduct(item.id);
+                           
+                            handleDeleteProduct(item._id);
                           }}
                         >
                           Delete
